Refactor Hyprland client adapter into modular components
