+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.3] - 2025-09-29
+
+### Changed
+
+- Reorganized the Hyprland client adapter into focused `config`, `sync_ops`,
+  and `listeners` modules, keeping the facade slim while re-exporting the
+  public surface.
+- Centralized retry and backoff utilities for synchronous requests and event
+  listeners to reuse.
+
+### Added
+
+- Unit tests covering the new retry delay helpers and listener backoff guard
+  paths.
+
 ## [0.6.2] - 2025-09-28
 
 ### Changed
